Added input validation
Added input validation for all fields in the form so the user will be prompted with an error message if there are any empty fields when trying to create a new task. Also added validation for the task duration so that the selected start time has to be before the selected finish time.